Extract resource code to a stand-alone package
